+++ .prettierignore
@@ -5,6 +5,6 @@ prisma/migrations/*
 prisma/reset.sh
 prisma/schema.prisma
 # Ignore Handlebars templates that use conditional attrs
-src/views/_partials/layout/head.html
-src/views/register.html
-src/views/project/blog/editor.html
+src/**/views/_partials/layout/head.html
+src/**/views/register.html
+src/**/views/project/blog/editor.html
