+++ src/core/handlers/projects/core/getAll.mjs
@@ -1,5 +1,5 @@
 import logger from "$/core/utils/logger.mjs";
-import prisma from "$/core/prisma.mjs";
+import prisma from "$/core/services/database.mjs";
 
 export default async function getAll(req, res) {
   try {
