+++ src/core/handlers/projects/core/remove.mjs
@@ -1,5 +1,5 @@
 import logger from "$/core/utils/logger.mjs";
-import prisma from "$/core/prisma.mjs";
+import prisma from "$/core/services/database.mjs";
 import {
   ensureProjectAccess,
   ProjectAccessError,
