+++ src/core/handlers/projects/index.mjs
@@ -4,7 +4,7 @@ import {
   disposeGitManager,
 } from "$/core/libs/git/registry.mjs";
 import logger from "$/core/utils/logger.mjs";
-import prisma from "$/core/prisma.mjs";
+import prisma from "$/core/services/database.mjs";
 import { ensureProjectAccess } from "$/core/utils/projectAccess.mjs";
 import { uuid } from "$/core/utils/id.mjs";
