+++ src/core/handlers/users/index.mjs
@@ -1,5 +1,5 @@
 import logger from "$/core/utils/logger.mjs";
-import prisma from "$/core/prisma.mjs";
+import prisma from "$/core/services/database.mjs";
 import { USER_ROLES } from "$/core/config/index.mjs";
 import { syncProjectPrimaryOwner } from "$/core/utils/projectAccess.mjs";
