+++ src/core/server.mjs
@@ -26,7 +26,10 @@ import hbsHelpers from "./utils/hbsHelpers.mjs";
 import logger from "./utils/logger.mjs";
 global.logger = logger; // Make logger globally accessible (e.g., in Prisma hooks)
 
-import { connectPrismaWithRetry, gracefulShutdown } from "./prisma.mjs";
+import {
+  connectPrismaWithRetry,
+  gracefulShutdown,
+} from "./services/database.mjs";
 import env from "./config/env.mjs";
 
 const { __publicdir, __templatedir, __datadir, PORT, NODE_ENV } = env();
