+++ tests/config-utils.test.mjs
@@ -1,7 +1,7 @@
 import test from "node:test";
 import assert from "node:assert/strict";
 
-import { toBool } from "$/config/utils.mjs";
+import { toBool } from "$/core/config/utils.mjs";
 
 test("toBool returns default for empty input", () => {
   assert.equal(toBool(undefined, true), true);
