+++ tests/projects-shares.test.mjs
@@ -1,8 +1,8 @@
 import test from "node:test";
 import assert from "node:assert/strict";
 
-import prisma from "$/prisma.mjs";
-import * as shares from "$/handlers/projects/shares.mjs";
+import prisma from "$/core/services/database.mjs";
+import * as shares from "$/core/handlers/projects/shares.mjs";
 
 function buildRes() {
   return {
